chore: Move naive-ui plugin to module
